+++ R/score.R
@@ -18,7 +18,7 @@
 #' set.seed(1)
 #' pts <- matrix(rnorm(60), ncol = 2)
 #' d <- dist(pts)
-#' MinDist(d, Gonzalez(d, 5L))
+#' MinDist(d, FarFirst(d, 5L))
 #' @export
 MinDist <- function(d = NULL, idx, points = NULL) {
   if (!is.null(points)) {